Changed NodeDB to SegmentDB, and made it only store the segments of nodes rather than all data related to nodes.
