+++ src/main/java/Parsing/ParserTest.java
@@ -1,10 +1,8 @@
 package parsing;
 
 import datastructure.Node;
-import datastructure.NodeDB;
+import datastructure.SegmentDB;
 import org.junit.jupiter.api.Test;
-
-import java.sql.ResultSet;
 
 import static org.junit.jupiter.api.Assertions.*;
 
@@ -21,7 +19,7 @@ class ParserTest {
     @Test
     void parse() {
         Parser parser = Parser.getInstance();
-        NodeDB data = parser.parse("/test2.gfa");
+        SegmentDB data = parser.parse("/test2.gfa");
         assertEquals("G",  data.getSegment(7));
         Node node = data.getNode(5);
         assertEquals(3, node.getIncomingEdges()[0]);
